+++ pymodbus/client/asynchronous/async_io/__init__.py
@@ -258,8 +258,7 @@ class ReconnectingAsyncioModbusTcpClient(object):
         """
         self.delay_ms = self.DELAY_MIN_MS
 
-    @asyncio.coroutine
-    def start(self, host, port=502):
+    async def start(self, host, port=502):
         """
         Initiates connection to start client
         :param host:
@@ -272,7 +271,7 @@ class ReconnectingAsyncioModbusTcpClient(object):
         _logger.debug('Connecting to %s:%s.' % (host, port))
         self.host = host
         self.port = port
-        yield from self._connect()
+        return await self._connect()
 
     def stop(self):
         """
@@ -295,13 +294,12 @@ class ReconnectingAsyncioModbusTcpClient(object):
         protocol.factory = self
         return protocol
 
-    @asyncio.coroutine
-    def _connect(self):
+    async def _connect(self):
         _logger.debug('Connecting.')
         try:
-            yield from self.loop.create_connection(self._create_protocol,
-                                                   self.host,
-                                                   self.port)
+            transport, protocol = await self.loop.create_connection(
+                self._create_protocol, self.host, self.port)
+            return transport, protocol
         except Exception as ex:
             _logger.warning('Failed to connect: %s' % ex)
             asyncio.ensure_future(self._reconnect(), loop=self.loop)
@@ -338,13 +336,12 @@ class ReconnectingAsyncioModbusTcpClient(object):
         else:
             _logger.error('Factory protocol disconnect callback called while not connected.')
 
-    @asyncio.coroutine
-    def _reconnect(self):
+    async def _reconnect(self):
         _logger.debug('Waiting %d ms before next '
                       'connection attempt.' % self.delay_ms)
-        yield from asyncio.sleep(self.delay_ms / 1000)
+        await asyncio.sleep(self.delay_ms / 1000)
         self.delay_ms = min(2 * self.delay_ms, self.DELAY_MAX_MS)
-        yield from self._connect()
+        return await self._connect()
 
 
 class AsyncioModbusTcpClient(object):
@@ -388,18 +385,17 @@ class AsyncioModbusTcpClient(object):
         protocol.factory = self
         return protocol
 
-    @asyncio.coroutine
-    def connect(self):
+    async def connect(self):
         """
         Connect and start Async client
         :return:
         """
         _logger.debug('Connecting.')
         try:
-            yield from self.loop.create_connection(self._create_protocol,
-                                                   self.host,
-                                                   self.port)
+            transport, protocol = await self.loop.create_connection(
+                self._create_protocol, self.host, self.port)
             _logger.info('Connected to %s:%s.' % (self.host, self.port))
+            return transport, protocol
         except Exception as ex:
             _logger.warning('Failed to connect: %s' % ex)
             # asyncio.asynchronous(self._reconnect(), loop=self.loop)
@@ -448,8 +444,7 @@ class ReconnectingAsyncioModbusTlsClient(ReconnectingAsyncioModbusTcpClient):
         self.framer = framer
         ReconnectingAsyncioModbusTcpClient.__init__(self, protocol_class, loop)
 
-    @asyncio.coroutine
-    def start(self, host, port=802, sslctx=None, server_hostname=None):
+    async def start(self, host, port=802, sslctx=None, server_hostname=None):
         """
         Initiates connection to start client
         :param host:
@@ -468,17 +463,17 @@ class ReconnectingAsyncioModbusTlsClient(ReconnectingAsyncioModbusTcpClient):
             self.sslctx.options |= ssl.OP_NO_SSLv3
             self.sslctx.options |= ssl.OP_NO_SSLv2
         self.server_hostname = server_hostname
-        yield from ReconnectingAsyncioModbusTcpClient.start(self, host, port)
+        return await ReconnectingAsyncioModbusTcpClient.start(self, host, port)
 
-    @asyncio.coroutine
-    def _connect(self):
+    async def _connect(self):
         _logger.debug('Connecting.')
         try:
-            yield from self.loop.create_connection(self._create_protocol,
-                                                   self.host,
-                                                   self.port,
-                                                   ssl=self.sslctx,
-                                                   server_hostname=self.server_hostname)
+            return await self.loop.create_connection(
+                self._create_protocol, self.host,
+                                      self.port,
+                                      ssl=self.sslctx,
+                                      server_hostname=self.server_hostname
+            )
         except Exception as ex:
             _logger.warning('Failed to connect: %s' % ex)
             asyncio.ensure_future(self._reconnect(), loop=self.loop)
@@ -531,8 +526,7 @@ class ReconnectingAsyncioModbusUdpClient(object):
         """
         self.delay_ms = 100
 
-    @asyncio.coroutine
-    def start(self, host, port=502):
+    async def start(self, host, port=502):
         """
         Start reconnecting asynchronous udp client
         :param host: Host IP to connect
@@ -548,12 +542,12 @@ class ReconnectingAsyncioModbusUdpClient(object):
         # - [(family, type, proto, canonname, sockaddr),]
         # We want sockaddr which is a (ip, port) tuple
         # udp needs ip addresses, not hostnames
-        addrinfo = yield from self.loop.getaddrinfo(host,
+        addrinfo = await self.loop.getaddrinfo(host,
                                                     port,
                                                     type=DGRAM_TYPE)
         self.host, self.port = addrinfo[0][-1]
 
-        yield from self._connect()
+        return await self._connect()
 
     def stop(self):
         """
@@ -578,17 +572,17 @@ class ReconnectingAsyncioModbusUdpClient(object):
         protocol.factory = self
         return protocol
 
-    @asyncio.coroutine
-    def _connect(self):
+    async def _connect(self):
         _logger.debug('Connecting.')
         try:
-            yield from self.loop.create_datagram_endpoint(
+            ep = await self.loop.create_datagram_endpoint(
                 functools.partial(self._create_protocol,
                                   host=self.host,
                                   port=self.port),
                 remote_addr=(self.host, self.port)
             )
             _logger.info('Connected to %s:%s.' % (self.host, self.port))
+            return ep
         except Exception as ex:
             _logger.warning('Failed to connect: %s' % ex)
             asyncio.ensure_future(self._reconnect(), loop=self.loop)
@@ -623,13 +617,12 @@ class ReconnectingAsyncioModbusUdpClient(object):
             _logger.error('Factory protocol disconnect '
                           'callback called while not connected.')
 
-    @asyncio.coroutine
-    def _reconnect(self):
+    async def _reconnect(self):
         _logger.debug('Waiting %d ms before next '
                       'connection attempt.' % self.delay_ms)
-        yield from asyncio.sleep(self.delay_ms / 1000)
+        await asyncio.sleep(self.delay_ms / 1000)
         self.delay_ms = min(2 * self.delay_ms, self.DELAY_MAX_MS)
-        yield from self._connect()
+        return await self._connect()
 
 
 class AsyncioModbusUdpClient(object):
@@ -680,21 +673,21 @@ class AsyncioModbusUdpClient(object):
         protocol.factory = self
         return protocol
 
-    @asyncio.coroutine
-    def connect(self):
+    async def connect(self):
         _logger.debug('Connecting.')
         try:
-            addrinfo = yield from self.loop.getaddrinfo(
+            addrinfo = await self.loop.getaddrinfo(
                 self.host,
                 self.port,
                 type=DGRAM_TYPE)
             _host, _port = addrinfo[0][-1]
-            yield from self.loop.create_datagram_endpoint(
+            ep = await self.loop.create_datagram_endpoint(
                 functools.partial(self._create_protocol,
                                   host=_host, port=_port),
                 remote_addr=(self.host, self.port)
             )
             _logger.info('Connected to %s:%s.' % (self.host, self.port))
+            return ep
         except Exception as ex:
             _logger.warning('Failed to connect: %s' % ex)
             # asyncio.asynchronous(self._reconnect(), loop=self.loop)
@@ -780,8 +773,7 @@ class AsyncioModbusSerialClient(object):
     def _connected(self):
         return self._connected_event.is_set()
 
-    @asyncio.coroutine
-    def connect(self):
+    async def connect(self):
         """
         Connect Async client
         :return:
@@ -790,11 +782,11 @@ class AsyncioModbusSerialClient(object):
         try:
             from serial_asyncio import create_serial_connection
 
-            yield from create_serial_connection(
+            await create_serial_connection(
                 self.loop, self._create_protocol, self.port, baudrate=self.baudrate,
                 bytesize=self.bytesize, stopbits=self.stopbits, parity=self.parity, **self._extra_serial_kwargs
             )
-            yield from self._connected_event.wait()
+            await self._connected_event.wait()
             _logger.info('Connected to %s', self.port)
         except Exception as ex:
             _logger.warning('Failed to connect: %s', ex)
@@ -830,8 +822,7 @@ class AsyncioModbusSerialClient(object):
                           'called while not connected.')
 
 
-@asyncio.coroutine
-def init_tcp_client(proto_cls, loop, host, port, **kwargs):
+async def init_tcp_client(proto_cls, loop, host, port, **kwargs):
     """
     Helper function to initialize tcp client
     :param proto_cls:
@@ -843,12 +834,11 @@ def init_tcp_client(proto_cls, loop, host, port, **kwargs):
     """
     client = ReconnectingAsyncioModbusTcpClient(protocol_class=proto_cls,
                                                 loop=loop)
-    yield from client.start(host, port)
+    await client.start(host, port)
     return client
 
 
-@asyncio.coroutine
-def init_tls_client(proto_cls, loop, host, port, sslctx=None,
+async def init_tls_client(proto_cls, loop, host, port, sslctx=None,
                     server_hostname=None, framer=None, **kwargs):
     """
     Helper function to initialize tcp client
@@ -864,12 +854,11 @@ def init_tls_client(proto_cls, loop, host, port, sslctx=None,
     """
     client = ReconnectingAsyncioModbusTlsClient(protocol_class=proto_cls,
                                                 loop=loop, framer=framer)
-    yield from client.start(host, port, sslctx, server_hostname)
+    await client.start(host, port, sslctx, server_hostname)
     return client
 
 
-@asyncio.coroutine
-def init_udp_client(proto_cls, loop, host, port, **kwargs):
+async def init_udp_client(proto_cls, loop, host, port, **kwargs):
     """
     Helper function to initialize UDP client
     :param proto_cls:
@@ -881,5 +870,5 @@ def init_udp_client(proto_cls, loop, host, port, **kwargs):
     """
     client = ReconnectingAsyncioModbusUdpClient(protocol_class=proto_cls,
                                                 loop=loop)
-    yield from client.start(host, port)
+    await client.start(host, port)
     return client